requirements.txt
asgiref==3.6.0
Django==4.1.4
django-modeltranslation==0.18.9
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
install==1.3.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
gunicorn==20.1.0
pillow==9.4.0
requests==2.28.2
googletrans==4.0.0-rc1
transliterate==1.10.2
drf-yasg==1.21.5